+++ tests/test_mailer.py
@@ -12,7 +12,7 @@ from django.core.exceptions import ImproperlyConfigured
 from django.core.management import call_command
 from django.test import TestCase
 from django.utils.timezone import now as datetime_now
-from mock import ANY, Mock, patch
+from mock import Mock, patch
 import six
 
 import mailer
@@ -90,20 +90,20 @@ class SendingTest(TestCase):
             self.assertEqual(Message.objects.count(), 0)
 
     def test_purge_old_entries(self):
-        # Send one successfully
-        with self.settings(MAILER_EMAIL_BACKEND="django.core.mail.backends.locmem.EmailBackend"):
-            mailer.send_mail("Subject", "Body", "sender1@example.com",
-                             ["recipient@example.com"])
-            engine.send_all()
 
-        # And one failure
-        with self.settings(MAILER_EMAIL_BACKEND="tests.FailingMailerEmailBackend"):
-            mailer.send_mail("Subject", "Body", "sender2@example.com",
-                             ["recipient@example.com"])
+        def send_mail(success):
+            backend = ("django.core.mail.backends.locmem.EmailBackend"
+                       if success else "tests.FailingMailerEmailBackend")
+            with self.settings(MAILER_EMAIL_BACKEND=backend):
+                mailer.send_mail("Subject", "Body", "sender@example.com", ["recipient@example.com"])
+                engine.send_all()
+                if not success:
+                    Message.objects.retry_deferred()
+                    engine.send_all()
 
-            engine.send_all()
-            Message.objects.retry_deferred()
-            engine.send_all()
+        # 1 success, 1 failure, and purge only success
+        send_mail(True)
+        send_mail(False)
 
         with patch.object(mailer.models, 'datetime_now') as datetime_now_patch:
             datetime_now_patch.return_value = datetime_now() + datetime.timedelta(days=2)
@@ -111,6 +111,25 @@ class SendingTest(TestCase):
 
         self.assertNotEqual(MessageLog.objects.filter(result=RESULT_FAILURE).count(), 0)
         self.assertEqual(MessageLog.objects.filter(result=RESULT_SUCCESS).count(), 0)
+
+        # 1 success, 1 failure, and purge only failures
+        send_mail(True)
+
+        with patch.object(mailer.models, 'datetime_now') as datetime_now_patch:
+            datetime_now_patch.return_value = datetime_now() + datetime.timedelta(days=2)
+            call_command('purge_mail_log', '1', '-r', 'failure')
+
+        self.assertEqual(MessageLog.objects.filter(result=RESULT_FAILURE).count(), 0)
+        self.assertNotEqual(MessageLog.objects.filter(result=RESULT_SUCCESS).count(), 0)
+
+        # 1 success, 1 failure, and purge everything
+        send_mail(False)
+
+        with patch.object(mailer.models, 'datetime_now') as datetime_now_patch:
+            datetime_now_patch.return_value = datetime_now() + datetime.timedelta(days=2)
+            call_command('purge_mail_log', '1', '-r', 'all')
+
+        self.assertEqual(MessageLog.objects.count(), 0)
 
     def test_send_loop(self):
         with self.settings(MAILER_EMAIL_BACKEND="django.core.mail.backends.locmem.EmailBackend"):
@@ -263,12 +282,12 @@ class SendingTest(TestCase):
             self.assertEqual(Message.objects.deferred().count(), 0)
 
         with self.settings(MAILER_EMAIL_BACKEND="tests.FailingMailerEmailBackend", MAILER_EMAIL_MAX_DEFERRED=2):  # noqa
-            # 2 will get deferred 3 remain undeferred
-            with patch("logging.warning") as w:
+            with patch('mailer.engine.logger.warning') as mock_warning:
+                # 2 will get deferred 3 remain undeferred
                 engine.send_all()
 
-                w.assert_called_once()
-                arg = w.call_args[0][0]
+                mock_warning.assert_called_once()
+                arg = mock_warning.call_args[0][0]
                 self.assertIn("EMAIL_MAX_DEFERRED", arg)
                 self.assertIn("stopping for this round", arg)
 
@@ -537,11 +556,11 @@ class MessagesTest(TestCase):
 
             msg.save()
 
-            with patch("logging.warning") as w:
+            with patch('mailer.engine.logger.warning') as mock_warning:
                 engine.send_all()
 
-                w.assert_called_once()
-                arg = w.call_args[0][0]
+                mock_warning.assert_called_once()
+                arg = mock_warning.call_args[0][0]
                 self.assertIn("message discarded due to failure in converting from DB", arg)
 
             self.assertEqual(Message.objects.count(), 0)
@@ -641,39 +660,32 @@ def call_command_with_cron_arg(command, cron_value):
         return call_command(command, cron=cron_value)
 
     # newer django; test parsing by passing argument as string
+    # --cron/c command option is deprecated
     return call_command(command, '--cron={}'.format(cron_value))
 
 
 class CommandHelperTest(TestCase):
     def test_send_mail_no_cron(self):
-        with patch('mailer.management.commands.send_mail.logging') as logging:
-            call_command('send_mail')
-            logging.basicConfig.assert_called_with(level=logging.DEBUG, format=ANY)
+        call_command('send_mail')
 
     def test_send_mail_cron_0(self):
-        with patch('mailer.management.commands.send_mail.logging') as logging:
-            call_command_with_cron_arg('send_mail', 0)
-            logging.basicConfig.assert_called_with(level=logging.DEBUG, format=ANY)
+        # deprecated
+        call_command_with_cron_arg('send_mail', 0)
 
     def test_send_mail_cron_1(self):
-        with patch('mailer.management.commands.send_mail.logging') as logging:
-            call_command_with_cron_arg('send_mail', 1)
-            logging.basicConfig.assert_called_with(level=logging.ERROR, format=ANY)
+        # deprecated
+        call_command_with_cron_arg('send_mail', 1)
 
     def test_retry_deferred_no_cron(self):
-        with patch('mailer.management.commands.retry_deferred.logging') as logging:
-            call_command('retry_deferred')
-            logging.basicConfig.assert_called_with(level=logging.DEBUG, format=ANY)
+        call_command('retry_deferred')
 
     def test_retry_deferred_cron_0(self):
-        with patch('mailer.management.commands.retry_deferred.logging') as logging:
-            call_command_with_cron_arg('retry_deferred', 0)
-            logging.basicConfig.assert_called_with(level=logging.DEBUG, format=ANY)
+        # deprecated
+        call_command_with_cron_arg('retry_deferred', 0)
 
     def test_retry_deferred_cron_1(self):
-        with patch('mailer.management.commands.retry_deferred.logging') as logging:
-            call_command_with_cron_arg('retry_deferred', 1)
-            logging.basicConfig.assert_called_with(level=logging.ERROR, format=ANY)
+        # deprecated
+        call_command_with_cron_arg('retry_deferred', 1)
 
 
 class EmailBackendSettingLoopTest(TestCase):
@@ -686,3 +698,32 @@ class EmailBackendSettingLoopTest(TestCase):
         self.assertIn('mailer.backend.DbBackend', str(catcher.exception))
         self.assertIn('EMAIL_BACKEND', str(catcher.exception))
         self.assertIn('MAILER_EMAIL_BACKEND', str(catcher.exception))
+
+
+class UseFileLockTest(TestCase):
+    """Test the MAILER_USE_FILE_LOCK setting."""
+
+    def setUp(self):
+        # mocking return_value to prevent "ValueError: not enough values to unpack"
+        self.patcher_acquire_lock = patch("mailer.engine.acquire_lock", return_value=(True, True))
+        self.patcher_release_lock = patch("mailer.engine.release_lock", return_value=(True, True))
+
+        self.mock_acquire_lock = self.patcher_acquire_lock.start()
+        self.mock_release_lock = self.patcher_release_lock.start()
+
+    def test_mailer_use_file_lock_enabled(self):
+        with self.settings(MAILER_EMAIL_BACKEND="django.core.mail.backends.locmem.EmailBackend"):
+            engine.send_all()
+        self.mock_acquire_lock.assert_called_once()
+        self.mock_release_lock.assert_called_once()
+
+    def test_mailer_use_file_lock_disabled(self):
+        with self.settings(MAILER_USE_FILE_LOCK=False,
+                           MAILER_EMAIL_BACKEND="django.core.mail.backends.locmem.EmailBackend"):
+            engine.send_all()
+        self.mock_acquire_lock.assert_not_called()
+        self.mock_release_lock.assert_not_called()
+
+    def tearDown(self):
+        self.patcher_acquire_lock.stop()
+        self.patcher_release_lock.stop()